bugfix and re-work in snmp_walk.py
(1) Bugfix, (1) Re-work to clean up output in 2.7 and 3

Bugfix:
[1] - Added omitted 'load_targets' function import on line 7 that resulted in traceback / crash when module ran.

Rework:
[1] - snmp_scans() has been changed to clean up the output when this def runs.  Prior to the re-work, the output
on error was quite ugly, chock full of traceback, printed errors, and more.  With this rework, the tracebacks are caught
and output is uniform in all cases (success and non-success).  

This re-work also in-lined the decode('utf') of 'results' on line 66 such that if additional functionality is added later, the results can be used as a string as needed, rather than the byte-encoded string returned by check_output normally, which often results in a TypeError being thrown.

As an added bonus, the re-work eliminated the need for the redundant check_results() function, since the test it was 
performing is now done internally to the snmp_scans().

Tested in both python 2.7 and python 3