Use relative imports for the lib module
Previously executing reconnoitre.py from a different
directory yielded the following error:

```
from lib.core.input import CliArgumentParser
ModuleNotFoundError: No module named 'lib.core'
```

This is because it was trying to import from lib, which is
not a python library (or in the pwd).

As per PEP-328, this change switches the imports to relative
imports which means that it works from an installed
location, but also maintains the previous support of
allowing local running.

Fixes codingo/Reconnoitre#118

Signed-off-by: Dave Walker (Daviey) <email@daviey.com>